majdnem kesz 5. feladat nem ment 8. eselytelen volt

+++ src/app/Fuvar.java
@@ -22,7 +22,7 @@ public class Fuvar {
         return ido;
     }
 
-    public int getOsszeg() {
+    public double getOsszeg() {
         return osszeg;
     }
